requirements.txt
# Core dependencies for Marketing Copy Generator
streamlit==1.28.1
requests==2.32.3
python-dotenv==1.0.1


